requirements.txt
msgraph-sdk
azure-identity
python-dotenv
